fix key mashup on tab switch
when switching between tabs, the available keys would get jumbled, often
in strange way which sometimes persist across several tab switchs or not
at all. The cause of this came from substle overlapping of activated
keys between tabs which were never deactivated (setting the key to
`false` in the Key mapping).

this commit changes the key activation process to explicitly activate
*and* deactivate keys on tab switches (and other UI movements, such as
with the template editor). Now within the *key.go specification for each
tab, *only* those keys which are needed for that tab are listed.